pyproject.toml
[project]
name = 'lastfm-charts'
version = '0.1.0'
requires-python = '>=3.11'
dependencies = [
    'python-dotenv==1.0.0',
    'pandas>=2.0.0',
    'requests>=2.31.0',
    'pyarrow>=14.0.0',
    'matplotlib>=3.7.0',
    'sqlalchemy>=2.0.0',
    'psycopg2-binary>=2.9.0',
    'tabulate>=0.9.0',
    'great-expectations>=1.0.0',
]